refactor: smart table types
