+++ projects/ion/src/lib/core/types/index.ts
@@ -18,6 +18,7 @@ export * from './notification';
 export * from './pagination';
 export * from './popconfirm';
 export * from './simple-menu';
+export * from './smart-table';
 export * from './status';
 export * from './tab-group';
 export * from './tooltip';
