+++ projects/ion/src/lib/core/types/smart-table.ts
@@ -1,0 +1,14 @@
+import { EventTable } from '../../table/utilsTable';
+import { SafeAny } from '../../utils/safe-any';
+import { PageEvent } from './pagination';
+
+export interface TableEvent {
+  event: EventTable;
+  rows_selected?: SafeAny[];
+  change_page?: PageEvent;
+  order?: {
+    column: string;
+    desc: boolean;
+  };
+  data?: SafeAny;
+}
